phase update 방식 수정, enemy 프리팹 구현 완료, phase data 입력 완료

+++ Assets/Scripts/Skill/Passive Skills/IncreaseAllStatsSkill.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class IncreaseAllStatsSkill : Skill
 {
     public float healthIncreaseAmount = 30f;
-    public float attackIncreaseAmount = 5f;
+    public int attackIncreaseAmount = 5;
     public float attackSpeedIncreaseAmount = 1.2f;
     public float moveSpeedIncreaseAmount = 1.2f;
